Move list of imperative verbs to a separate file
