+++ src/legacy-v1/utilities/IterableUtilities.ts
@@ -1,22 +1,3 @@
-export function indexOfFromBinarySearch(sortedValues: Array<string>, target: string): number {
-	function binarySearch(inclusiveStartIndex: number, inclusiveEndIndex: number): number {
-		const middleIndex = Math.floor((inclusiveStartIndex + inclusiveEndIndex) / 2)
-		const comparison = target.localeCompare(sortedValues[middleIndex] ?? "", "en")
-
-		if (comparison === 0) {
-			return middleIndex
-		}
-		if (inclusiveStartIndex >= inclusiveEndIndex) {
-			return -1
-		}
-		return comparison < 0
-			? binarySearch(inclusiveStartIndex, middleIndex - 1)
-			: binarySearch(middleIndex + 1, inclusiveEndIndex)
-	}
-
-	return binarySearch(0, sortedValues.length - 1)
-}
-
 export function countOccurrences(value: string, characterToCount: string): number {
 	return value.split(characterToCount).length - 1
 }
